# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/flows": {
      "post": {
        "description": "Create a new workflow with steps and configuration",
        "operationId": "FlowsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Flow definition",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Flow name",
                    "example": "User Onboarding Flow"
                  },
                  "description": {
                    "type": "string",
                    "description": "Flow description",
                    "example": "Automated user onboarding process with email and notifications"
                  },
                  "version": {
                    "type": "number",
                    "description": "Flow version number",
                    "example": 1
                  },
                  "steps": {
                    "type": "array",
                    "description": "Workflow steps definition",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "step_1"
                        },
                        "type": {
                          "type": "string",
                          "example": "action"
                        },
                        "actionId": {
                          "type": "string",
                          "example": "action_abc123"
                        },
                        "config": {
                          "type": "object",
                          "example": {
                            "timeout": 30000
                          }
                        }
                      }
                    },
                    "example": [
                      {
                        "id": "step_1",
                        "type": "action",
                        "actionId": "action_send_email",
                        "config": {
                          "to": "{{user.email}}",
                          "template": "welcome"
                        }
                      }
                    ]
                  },
                  "settings": {
                    "type": "object",
                    "description": "Flow execution settings",
                    "example": {
                      "timeout": 300000,
                      "retries": 2
                    }
                  }
                },
                "required": [
                  "name",
                  "steps"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Flow created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "flow_abc123"
                    },
                    "name": {
                      "type": "string",
                      "example": "User Onboarding Flow"
                    },
                    "description": {
                      "type": "string",
                      "example": "Automated user onboarding process"
                    },
                    "version": {
                      "type": "number",
                      "example": 1
                    },
                    "steps": {
                      "type": "array",
                      "example": []
                    },
                    "settings": {
                      "type": "object",
                      "example": {
                        "timeout": 300000
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid flow definition"
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "Create Flow",
        "tags": [
          "Flows"
        ]
      },
      "get": {
        "description": "Get all workflows for the organization",
        "operationId": "FlowsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved flows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "flow_abc123"
                      },
                      "name": {
                        "type": "string",
                        "example": "User Onboarding Flow"
                      },
                      "version": {
                        "type": "number",
                        "example": 1
                      },
                      "steps": {
                        "type": "array",
                        "example": []
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "List Flows",
        "tags": [
          "Flows"
        ]
      }
    },
    "/flows/{id}": {
      "get": {
        "description": "Get a specific workflow by ID",
        "operationId": "FlowsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Flow ID",
            "schema": {
              "example": "flow_abc123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved flow"
          },
          "404": {
            "description": "Flow not found"
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "Get Flow",
        "tags": [
          "Flows"
        ]
      },
      "put": {
        "description": "Update workflow definition and configuration",
        "operationId": "FlowsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Flow ID",
            "schema": {
              "example": "flow_abc123",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Updated flow definition",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Flow name",
                    "example": "Updated User Onboarding Flow"
                  },
                  "description": {
                    "type": "string",
                    "description": "Flow description",
                    "example": "Enhanced automated user onboarding process"
                  },
                  "version": {
                    "type": "number",
                    "description": "Flow version number",
                    "example": 2
                  },
                  "steps": {
                    "type": "array",
                    "description": "Updated workflow steps definition",
                    "example": [
                      {
                        "id": "step_1",
                        "type": "action",
                        "actionId": "action_send_email",
                        "config": {
                          "to": "{{user.email}}",
                          "template": "welcome_v2"
                        }
                      }
                    ]
                  },
                  "settings": {
                    "type": "object",
                    "description": "Updated flow execution settings",
                    "example": {
                      "timeout": 600000,
                      "retries": 3
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Flow updated successfully"
          },
          "404": {
            "description": "Flow not found"
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "Update Flow",
        "tags": [
          "Flows"
        ]
      },
      "delete": {
        "description": "Delete a workflow permanently",
        "operationId": "FlowsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Flow ID",
            "schema": {
              "example": "flow_abc123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Flow deleted successfully"
          },
          "404": {
            "description": "Flow not found"
          },
          "409": {
            "description": "Flow has active executions and cannot be deleted"
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "Delete Flow",
        "tags": [
          "Flows"
        ]
      }
    },
    "/flows/{id}/execute": {
      "post": {
        "description": "Execute a workflow either synchronously or asynchronously with optional input variables",
        "operationId": "FlowsController_executeFlow",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Flow ID to execute",
            "schema": {
              "example": "flow_abc123",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Execution configuration",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "variables": {
                    "type": "object",
                    "description": "Input variables for the flow execution",
                    "example": {
                      "userId": "user_123",
                      "email": "user@example.com"
                    }
                  },
                  "useDurable": {
                    "type": "boolean",
                    "description": "Whether to use durable (async) execution",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Flow execution started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "executionId": {
                      "type": "string",
                      "example": "exec_abc123"
                    },
                    "status": {
                      "type": "string",
                      "example": "running"
                    },
                    "mode": {
                      "type": "string",
                      "example": "durable"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid execution parameters"
          },
          "404": {
            "description": "Flow not found"
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "Execute Flow",
        "tags": [
          "Flows"
        ]
      }
    },
    "/flows/{id}/executions": {
      "get": {
        "description": "Get execution history for a specific flow",
        "operationId": "FlowsController_getFlowExecutions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Flow ID",
            "schema": {
              "example": "flow_abc123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved flow executions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "executionId": {
                        "type": "string",
                        "example": "exec_abc123"
                      },
                      "status": {
                        "type": "string",
                        "example": "completed"
                      },
                      "startedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "completedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "variables": {
                        "type": "object",
                        "example": {
                          "userId": "user_123"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Flow not found"
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "Get Flow Executions",
        "tags": [
          "Flows"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tolstoy API",
    "description": "Interactive API reference for Tolstoy workflow automation platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://tolstoy.getpullse.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "x-org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-org-id",
        "description": "Organization ID for multi-tenant access"
      },
      "x-user-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "User ID for request context"
      }
    },
...
  }
}